+++ lib/api/sfApi.js
@@ -740,7 +740,7 @@ async function readAccountTemplateById(
 ) {
   apiUtils.setAxiosDefaults(firmId);
   try {
-    const response = await axios.get(`export_files/${accountTemplateId}`);
+    const response = await axios.get(`account_templates/${accountTemplateId}`);
     apiUtils.responseSuccessHandler(response);
     return response.data;
   } catch (error) {
@@ -772,7 +772,7 @@ async function findAccountTemplateByName(
     return;
   }
   const accountTemplate = accountTemplates.find(
-    (element) => element["name"] === accountTemplateName
+    (element) => element["name_nl"] === accountTemplateName
   );
   if (accountTemplate) {
     return await readAccountTemplateById(firmId, accountTemplate.id);
