+++ lib/templates/accountTemplate.js
@@ -1,10 +1,18 @@
 const fs = require("fs");
-const fsUtils = require("./utils/fsUtils");
-const templateUtils = require("./utils/templateUtils");
+const fsUtils = require("../utils/fsUtils");
+const templateUtils = require("../utils/templateUtils");
 
-class ExportFile {
-  static CONFIG_ITEMS = ["name", "file_name", "externally_managed", "encoding"];
-  static TEMPLATE_TYPE = "exportFile";
+class AccountTemplate {
+  static CONFIG_ITEMS = [
+    "name_en",
+    "name_fr",
+    "name_nl",
+    "text_configuration",
+    "externally_managed",
+    "account_range",
+    "mapping_list_ranges",
+  ];
+  static TEMPLATE_TYPE = "accountTemplate";
   static TEMPLATE_FOLDER = fsUtils.FOLDERS[this.TEMPLATE_TYPE];
   constructor() {}
 
@@ -15,15 +23,19 @@ class ExportFile {
    */
   static save(firmId, template) {
     if (templateUtils.missingLiquidCode(template)) return;
-    if (!templateUtils.checkValidName(template.name)) return;
+    if (!templateUtils.checkValidName(template.name_nl)) return; //TODO: Replace name_nl
 
-    const name = template.name;
+    const name = template.name_nl; //TODO: Replace name_nl
     fsUtils.createTemplateFolders(this.TEMPLATE_TYPE, name);
 
     // Liquid files
     const mainPart = template.text;
     const textParts = templateUtils.filterParts(template);
     fsUtils.createTemplateFiles(this.TEMPLATE_TYPE, name, mainPart, textParts);
+
+    // Liquid Test YAML
+    let testContent = "# Add your Liquid Tests here";
+    fsUtils.createLiquidTestFiles(this.TEMPLATE_TYPE, handle, testContent);
 
     // Config Json File
     let existingConfig = fsUtils.readConfig(this.TEMPLATE_TYPE, name);
@@ -33,6 +45,7 @@ class ExportFile {
         ...existingConfig?.id,
         [firmId]: template.id,
       },
+      test: `tests/${name}_liquid_test.yml`,
       ...configDetails,
     };
     fsUtils.writeConfig(this.TEMPLATE_TYPE, name, configContent);
@@ -114,21 +127,22 @@ class ExportFile {
   }
 }
 
-module.exports = { ExportFile };
+module.exports = { AccountTemplate };
 
 // Example Silverfin Response
 // {
-//   "id": 1,
-//   "marketplace_template_id": 10,
-//   "name": "New Export file",
-//   "file_name": "export_file_name.sxbrl",
-//   "externally_managed": false,
-//   "encoding": "UTF-8",
+//   "id": 100,
+//   "marketplace_template_id": 25,
+//   "account_range": null,
+//   "text_configuration": null,
+//   "name_en": Account template Name,
 //   "text": "Liquid code",
 //   "text_parts": [
 //       {
 //           "name": "part_1",
 //           "content": "part 1 liquid code"
 //       }
-//   ]
+//   ],
+//   "mapping_list_ranges": [],
+//   "externally_managed": false
 // }
