+++ lib/utils/apiUtils.js
@@ -109,6 +109,7 @@ async function responseErrorHandler(
   callbackParameters
 ) {
   if (error && error.response) {
+    
     consola.error(
       `Response Status: ${error.response.status} (${error.response.statusText}) - method: ${error.response.config.method} - url: ${error.response.config.url}`
     );
