+++ lib/utils/fsUtils.js
@@ -17,7 +17,6 @@ function createFolder(path) {
 }
 
 function createTemplateFolders(templateType, handle, testFolder = true) {
-  createFolder(path.join(process.cwd(), `${FOLDERS[templateType]}`));
   createFolder(
     path.join(process.cwd(), `${FOLDERS[templateType]}`, `${handle}`)
   );
@@ -176,7 +175,7 @@ function createConfigIfMissing(templateType, handle) {
 
   const config = {};
   config.id = {};
-  config.externally_managed = false;
+  config.externally_managed = true;
   switch (templateType) {
     case "reconciliationText":
       config.handle = handle;
@@ -210,6 +209,9 @@ function createConfigIfMissing(templateType, handle) {
       config.mapping_list_ranges = [];
       config.text = "main.liquid";
       config.text_parts = {};
+      config.name_nl = `${handle}`
+      config.name_fr = `${handle}`
+      config.name_en = `${handle}`
       break;
   }
   writeConfig(templateType, handle, config);
